+++ documentation/examples/custom-action-typescript/text-run/hello-world.ts
@@ -1,0 +1,5 @@
+import { ActionArgs } from "../../../.."
+
+export default function HelloWorld(action: ActionArgs) {
+  action.log("Hello World from TypeScript!")
+}